refactor: consolidate argument parsing to cmd_arg_parse.h and remove client-specific version
