+++ Client/examples/client_lib_writer_example.cpp
@@ -1,5 +1,5 @@
 #include "ClientConfiguration.h"
-#include "client_cmd_arg_parse.h"
+#include "cmd_arg_parse.h"
 #include <chronolog_client.h>
 #include "chrono_monitor.h"
 #include <common.h>
@@ -7,7 +7,7 @@
 
 int main(int argc, char** argv) {
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
