+++ test/integration/Client/client_lib_connect_rpc_test.cpp
@@ -2,9 +2,9 @@
 #include <unordered_map>
 #include <cassert>
 
+#include <cmd_arg_parse.h>
 #include "chronolog_client.h"
 #include "ClientConfiguration.h"
-#include "client_cmd_arg_parse.h"
 #include "chrono_monitor.h"
 #include "common.h"
 
@@ -13,7 +13,7 @@
 int main(int argc, char** argv)
 {
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
