+++ test/integration/Client/client_lib_hybrid_argobots_test.cpp
@@ -8,7 +8,6 @@
 #include <cmd_arg_parse.h>
 #include "chrono_monitor.h"
 #include "ClientConfiguration.h"
-#include "client_cmd_arg_parse.h"
 
 chronolog::Client*client;
 
@@ -54,7 +53,7 @@ int main(int argc, char**argv)
     MPI_Init_thread(&argc, &argv, MPI_THREAD_MULTIPLE, &provided);
 
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
