+++ test/integration/Client/client_lib_metadata_rpc_test.cpp
@@ -7,7 +7,7 @@
 #include "chronolog_client.h"
 #include "common.h"
 #include "ClientConfiguration.h"
-#include "client_cmd_arg_parse.h"
+#include "cmd_arg_parse.h"
 #include "chrono_monitor.h"
 
 #define NUM_CHRONICLE 10
@@ -18,7 +18,7 @@
 int main(int argc, char** argv) {
     
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
