+++ test/integration/Client/client_lib_multi_argobots_test.cpp
@@ -6,7 +6,6 @@
 #include <cmd_arg_parse.h>
 #include "ClientConfiguration.h"
 #include "chrono_monitor.h"
-#include "client_cmd_arg_parse.h"
 
 #define CHRONICLE_NAME_LEN 32
 #define STORY_NAME_LEN 32
@@ -82,7 +81,7 @@ int main(int argc, char** argv) {
     std::vector<std::thread> thread_vec;
 
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
