+++ test/integration/Client/client_lib_multi_openmp_test.cpp
@@ -5,7 +5,6 @@
 #include <omp.h>
 #include <cmd_arg_parse.h>
 #include "chrono_monitor.h"
-#include "client_cmd_arg_parse.h"
 #include "ClientConfiguration.h"
 
 #define STORY_NAME_LEN 32
@@ -13,7 +12,7 @@
 int main(int argc, char**argv)
 {
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
