+++ test/integration/Client/client_lib_story_reader.cpp
@@ -5,7 +5,6 @@
 #include <chrono>
 #include <cmd_arg_parse.h>
 #include "chrono_monitor.h"
-#include "client_cmd_arg_parse.h"
 #include "ClientConfiguration.h"
 
 #define STORY_NAME_LEN 5
@@ -136,7 +135,7 @@ void reader_thread( int tid, struct thread_arg * t)
 int main(int argc, char**argv)
 {
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
