+++ test/integration/Client/client_lib_thread_interdependency_test.cpp
@@ -12,7 +12,6 @@
 #include <algorithm>
 #include <iostream>
 #include "ClientConfiguration.h"
-#include "client_cmd_arg_parse.h"
 
 #define STORY_NAME_LEN 5
 #define DEFAULT_NUM_THREADS 4
@@ -431,7 +430,7 @@ int parse_num_threads_arg(int argc, char**argv)
 int main(int argc, char**argv)
 {
     // Load configuration
-    std::string conf_file_path = chronolog::parse_conf_path_arg(argc, argv);
+    std::string conf_file_path = parse_conf_path_arg(argc, argv);
     chronolog::ClientConfiguration confManager;
     if (!conf_file_path.empty()) {
         if (!confManager.load_from_file(conf_file_path)) {
